This puppet module currently does the following: - ensures that flume-repose & cf-flume-sink rpms exist or removes them - creates the /etc/sysconfig/flume-ng - contains spec tests
One note:  This module does not install any rpms, unlike puppet-repose.
When I approached @gregswift with a question on how to incorporate 2 distinct
version numbers into a single puppet module, he mentioned that he actually
prefers to manage it outside of puppet.  I was fine wtih that as eventually
the Ops group will be deploying this with other groups.

Future PR's will flesh out the actual flume configuration.

@gregswift - what's the best way to test this?  Can we configure my inova
to pull down this conf and I can verify that it works?